AsyncCredentialBase -> AsyncContextManager
