+++ sdk/identity/azure-identity/azure/identity/aio/_credentials/azure_cli.py
@@ -7,7 +7,7 @@ import sys
 import os
 
 from azure.core.exceptions import ClientAuthenticationError
-from .._credentials.base import AsyncCredentialBase
+from .._internal import AsyncContextManager
 from .._internal.decorators import log_get_token_async
 from ... import CredentialUnavailableError
 from ..._credentials.azure_cli import (
@@ -22,7 +22,7 @@ from ..._credentials.azure_cli import (
 from ..._internal import _scopes_to_resource
 
 
-class AzureCliCredential(AsyncCredentialBase):
+class AzureCliCredential(AsyncContextManager):
     """Authenticates by requesting a token from the Azure CLI.
 
     This requires previously logging in to Azure via "az login", and will use the CLI's currently logged in identity.
